Create a common directory
