+++ src/index.ts
@@ -2,5 +2,5 @@
 // Distributed under the terms of the Modified BSD License.
 'use strict';
 
-export * from './kernel-selector/index';
+export * from './common/selectkernel';
 export * from './notebook/index';
