+++ src/notebook/index.ts
@@ -2,13 +2,14 @@
 // Distributed under the terms of the Modified BSD License.
 'use strict';
 
+export * from './actions';
+export * from './default-toolbar';
 export * from './model';
+export * from './modelfactory';
 export * from './nbformat';
+export * from './panel';
+export * from './toolbar';
 export * from './trust';
 export * from './widget';
-export * from './toolbar';
-export * from './default-toolbar';
-export * from './actions';
-export * from './modelfactory';
 export * from './widgetfactory';
-export * from './panel';
+
